refactor(chunking): move document chunking into its own component
Chunking sits between sources and vector stores in the ingestion
pipeline, so it now lives at components/chunking/ instead of inside
vector_stores/: chunker.py (facade + shared helpers), docling_local,
docling_remote, docling_serve_client.
