+++ backend/syft_space/components/chunking/docling_remote.py
@@ -13,12 +13,12 @@ from collections.abc import Callable
 from pathlib import Path
 from typing import Any
 
-from syft_space.components.shared.ingest_types import IngestFile
-from syft_space.components.vector_stores.chunking import (
+from syft_space.components.chunking.chunker import (
     _extract_pictures_and_chunks,
     _get_pdf_page_count,
 )
-from syft_space.components.vector_stores.docling_serve_client import DoclingServeClient
+from syft_space.components.chunking.docling_serve_client import DoclingServeClient
+from syft_space.components.shared.ingest_types import IngestFile
 
 logger = logging.getLogger(__name__)
